+++ components/progress-ring/__docs__/storybook-stories.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { PROGRESS_RING } from '../../../utilities/constants';
+
+import Active from '../__examples__/active';
 import Base from '../__examples__/base';
 import Complete from '../__examples__/complete';
 import Warning from '../__examples__/warning';
@@ -8,13 +9,16 @@ import Expired from '../__examples__/expired';
 import CustomIcon from '../__examples__/customIcon';
 import Examples from '../__examples__/examples';
 
+import { PROGRESS_RING } from '../../../utilities/constants';
+
 storiesOf(PROGRESS_RING, module)
 	.addDecorator((getStory) => (
 		<div className="slds-p-around_medium">{getStory()}</div>
 	))
 	.add('Base', () => <Base />)
+	.add('Theme: Active', () => <Active />)
 	.add('Theme: Complete', () => <Complete />)
 	.add('Theme: Warning', () => <Warning />)
 	.add('Theme: Expired', () => <Expired />)
 	.add('Custom Icon', () => <CustomIcon />)
-	.add('Docs site Examples', () => <Examples />);
+	.add('Snapshot Test Examples', () => <Examples />);