+++ components/radio-group/index.jsx
@@ -15,6 +15,13 @@ import { RADIO_GROUP } from '../../utilities/constants';
 
 const propTypes = {
 	/**
+	 * **Assistive text for accessibility**
+	 * * `label`: This label appears in the legend.
+	 */
+	assistiveText: PropTypes.shape({
+		label: PropTypes.string,
+	}),
+	/**
 	 * Children are expected to be Radio components.
 	 */
 	children: PropTypes.node.isRequired,
@@ -56,25 +63,22 @@ const propTypes = {
 	 * The ID of the error message, for linking to radio inputs with aria-describedby.
 	 */
 	errorId: PropTypes.string,
+	/**
+	 * Variants of radio groups such as Radio Button Group
+	 */
+	variant: PropTypes.oneOf(['base', 'button-group']),
 };
 
-const defaultProps = { labels: {} };
+const defaultProps = { assistiveText: {}, labels: {}, variant: 'base' };
 
 /**
  * A styled select list that can have a single entry checked at any one time.
  * The RadioGroup component wraps [Radio](/components/radios) components, which should be used as children.
  */
 class RadioGroup extends React.Component {
-	constructor(props) {
-		super(props);
-
-		// Merge objects of strings with their default object
-		this.labels = this.props.labels
-			? assign({}, defaultProps.labels, this.props.labels)
-			: defaultProps.labels;
-
+	componentWillMount() {
 		this.generatedName = shortid.generate();
-		this.generatedErrorId = this.labels.error ? shortid.generate() : null;
+		this.generatedErrorId = shortid.generate();
 	}
 
 	getErrorId() {
@@ -93,6 +97,11 @@ class RadioGroup extends React.Component {
 	}
 
 	render() {
+		// Merge objects of strings with their default object
+		this.labels = this.props.labels
+			? assign({}, defaultProps.labels, this.props.labels)
+			: defaultProps.labels;
+
 		const children = React.Children.map(this.props.children, (child) =>
 			React.cloneElement(child, {
 				name: this.getName(),
@@ -108,13 +117,21 @@ class RadioGroup extends React.Component {
 					'slds-has-error': this.labels.error,
 				})}
 			>
-				<legend className="slds-form-element__legend slds-form-element__label">
+				<legend
+					className={classNames(
+						'slds-form-element__legend',
+						'slds-form-element__label',
+						this.props.assistiveText.label ? 'slds-assistive-text' : ''
+					)}
+				>
 					{this.props.required ? (
 						<abbr className="slds-required" title="required">
 							*
 						</abbr>
 					) : null}
-					{this.labels.label}
+					{this.props.assistiveText.label
+						? this.props.assistiveText.label
+						: this.labels.label}
 				</legend>
 				<div
 					className={classNames(
@@ -122,7 +139,14 @@ class RadioGroup extends React.Component {
 						this.props.className
 					)}
 				>
-					{children}
+					{this.props.variant === 'button-group' ? (
+						<div style={this.props.style} className="slds-radio_button-group">
+							{children}
+						</div>
+					) : (
+						children
+					)}
+
 					{this.labels.error ? (
 						<div id={this.getErrorId()} className="slds-form-element__help">
 							{this.labels.error}