+++ RELEASENOTES.md
@@ -31,9 +31,38 @@ These are changes that have backwards-compatible solutions present and that comp
 
 ### Latest Release
 
+## Release 0.10.5
+
+**Major Features**
+
+* `Setup Assistant`: Add Component
+
+**Minor Features**
+
+* `ProgressRing`: Add active theme
+* `ButtonGroup`: Adds `list` variant
+* `Modal`: Add `small` size support
+
+**Bugfixes**
+
+* Radio Group: Update labels from props after instance construction. Label and error label did not update after component is initialzied.
+* `SetupAssistant`: Design review and red-lining of component such as padding, spacing, and borders.
+* `ColorPicker`: Removes some differences between SLDS `ColorPicker` pattern and this library
+
+**Documentation**
+
+* Codebase: Add default prop change warning
+* `SLDSAccordian`: 3 dropdowns in base whereas only 2 in base-open
+* `Combobox`: Dialog variant documation site example fix
+
+**Maintainance**
+
+* Merge Radio Button Group and Radio Group, so that one is composed over the other.
+
 ## Release 0.10.4
 
 **Major Features**
+
 * `BuilderHeader`: Add component
 * `Expandable Section`: Add component
 
@@ -46,11 +75,11 @@ These are changes that have backwards-compatible solutions present and that comp
 * `ProgressIndicator`: Adds `setup-assistant` variant
 * `Modal`: Refactor `border-radius` render logic based on header and footer
 * `GlobalNavigation`: Align closer to SLDS
-* `AppLauncher`: Align closer to SLDS 
+* `AppLauncher`: Align closer to SLDS
 
 **Bugfixes**
 
-* `Carousel`: Prevent focus grab when slide changes 
+* `Carousel`: Prevent focus grab when slide changes
 * `DataTable`: Fixed header scrolling bugfix when `tab` is used
 * `Illustration`: Minor bugfix for `style` object
 
@@ -60,7 +89,6 @@ These are changes that have backwards-compatible solutions present and that comp
 * Fix accessibility issues in doc site examples
 * `GlobalNavigation`: Add a11y exception to region
 * Sync up the `package-lock.json` file
-
 
 ## Release 0.10.3